+++ sigma-tree/src/chain/transaction.rs
@@ -1,5 +1,7 @@
 //! Ergo transaction
 
+pub mod unsigned;
+
 #[cfg(feature = "with-serde")]
 use super::json;
 use super::{
@@ -9,8 +11,6 @@ use super::{
     input::Input,
     token::TokenId,
     ErgoBox,
-    ProofBytes::Empty,
-    ProverResult, UnsignedInput,
 };
 use crate::{
     eval::Env,
@@ -31,6 +31,7 @@ use std::io;
 use std::iter::FromIterator;
 #[cfg(feature = "with-serde")]
 use thiserror::Error;
+use unsigned::UnsignedTransaction;
 
 /// Transaction id (ModifierId in sigmastate)
 #[derive(PartialEq, Eq, Hash, Debug, Clone)]
@@ -59,67 +60,6 @@ impl SigmaSerializable for TxId {
 impl Into<String> for TxId {
     fn into(self) -> String {
         self.0.into()
-    }
-}
-
-/// Unsigned (inputs without proofs) transaction
-#[derive(PartialEq, Debug, Clone)]
-pub struct UnsignedTransaction {
-    tx_id: TxId,
-    /// unsigned inputs, that will be spent by this transaction.
-    pub inputs: Vec<UnsignedInput>,
-    /// inputs, that are not going to be spent by transaction, but will be reachable from inputs
-    /// scripts. `dataInputs` scripts will not be executed, thus their scripts costs are not
-    /// included in transaction cost and they do not contain spending proofs.
-    pub data_inputs: Vec<DataInput>,
-    /// box candidates to be created by this transaction
-    pub output_candidates: Vec<ErgoBoxCandidate>,
-}
-
-impl UnsignedTransaction {
-    /// Creates new transation
-    pub fn new(
-        inputs: Vec<UnsignedInput>,
-        data_inputs: Vec<DataInput>,
-        output_candidates: Vec<ErgoBoxCandidate>,
-    ) -> UnsignedTransaction {
-        let tx_to_sign = UnsignedTransaction {
-            tx_id: TxId::zero(),
-            inputs,
-            data_inputs,
-            output_candidates,
-        };
-        let tx_id = tx_to_sign.calc_tx_id();
-        UnsignedTransaction {
-            tx_id,
-            ..tx_to_sign
-        }
-    }
-
-    fn calc_tx_id(&self) -> TxId {
-        let bytes = self.bytes_to_sign();
-        TxId(blake2b256_hash(&bytes))
-    }
-
-    /// message to be signed by the [`Prover`] (serialized tx)
-    pub fn bytes_to_sign(&self) -> Vec<u8> {
-        let empty_proofs_input = self
-            .inputs
-            .iter()
-            .map(|ui| Input {
-                box_id: ui.box_id.clone(),
-                spending_proof: ProverResult {
-                    proof: Empty,
-                    extension: ui.extension.clone(),
-                },
-            })
-            .collect();
-        let tx = Transaction::new(
-            empty_proofs_input,
-            self.data_inputs.clone(),
-            self.output_candidates.clone(),
-        );
-        tx.sigma_serialise_bytes()
     }
 }
 
@@ -406,21 +346,6 @@ pub mod tests {
         type Strategy = BoxedStrategy<Self>;
     }
 
-    impl Arbitrary for UnsignedTransaction {
-        type Parameters = ();
-
-        fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
-            (
-                vec(any::<UnsignedInput>(), 1..10),
-                vec(any::<DataInput>(), 0..10),
-                vec(any::<ErgoBoxCandidate>(), 1..10),
-            )
-                .prop_map(|(inputs, data_inputs, outputs)| Self::new(inputs, data_inputs, outputs))
-                .boxed()
-        }
-        type Strategy = BoxedStrategy<Self>;
-    }
-
     proptest! {
 
         #[test]
@@ -432,11 +357,6 @@ pub mod tests {
         #[test]
         fn tx_id_ser_roundtrip(v in any::<TxId>()) {
             prop_assert_eq![sigma_serialize_roundtrip(&v), v];
-        }
-
-        #[test]
-        fn test_unsigned_tx_bytes_to_sign(v in any::<UnsignedTransaction>()) {
-            prop_assert!(!v.bytes_to_sign().is_empty());
         }
 
         // #[test]
